move message types to core
