+++ packages/wallet-sdk/src/core/message/ConfigMessage.ts
@@ -1,4 +1,4 @@
-import { Message } from ':core/communicator/Message';
+import { Message } from ':core/message/Message';
 
 export interface ConfigMessage extends Message {
   type: 'config';
