+++ packages/wallet-sdk/src/core/message/Message.ts
@@ -3,4 +3,5 @@ import { UUID } from 'crypto';
 export interface Message {
   type: 'config' | 'scw';
   id: UUID;
+  version: string;
 }
