+++ packages/wallet-sdk/src/sign/scw/transport/PopUpCommunicator.ts
@@ -1,10 +1,10 @@
 import { UUID } from 'crypto';
 
-import { ClientConfigEventType, isConfigMessage, SignerType } from './ConfigMessage';
 import { PopUpConfigurator } from './PopUpConfigurator';
 import { CrossDomainCommunicator } from ':core/communicator/CrossDomainCommunicator';
-import { Message } from ':core/communicator/Message';
 import { standardErrors } from ':core/error';
+import { ClientConfigEventType, isConfigMessage, SignerType } from ':core/message/ConfigMessage';
+import { Message } from ':core/message/Message';
 
 // TODO: how to set/change configurations?
 const POPUP_WIDTH = 420;
